requirements.txt
keyboard==0.13.5
python-dotenv==1.0.1
